extract activate agent
